Create data invalidation factory
