+++ src/core/pipe/pipe.invalidate.factory.js
@@ -1,0 +1,16 @@
+import Middleware from './pipe.middleware';
+
+export function pipeInvalidateFactory(model) {
+	return function (source, diff) {
+		const stages = model.data().pipe;
+		const middleware = new Middleware(stages);
+		const context = {
+			model: model,
+			source: source,
+			diff: diff
+		};
+		middleware
+			.run(context, model.data().rows)
+			.then(result => model.data({rows: result}));
+	};
+}
